Connexion ok sans hachage

+++ escalade/escalade-modele/src/main/java/fr/escalade/beans/Utilisateur.java
@@ -48,7 +48,7 @@ public class Utilisateur implements Serializable {
 	public long getIduser() {
 		return iduser;
 	}
-	public void setId_user(int iduser) {
+	public void setId_user(Long iduser) {
 		this.iduser = iduser;
 	}
